Moved Slic test to Slic transport tests
