+++ tests/IceRpc.Conformance.Tests/Transports/MultiplexedConnectionConformanceTests.cs
@@ -359,89 +359,6 @@ public abstract class MultiplexedConnectionConformanceTests
         Assert.That(async () => await connectTask, Throws.InstanceOf<OperationCanceledException>());
     }
 
-    /// <summary>Verifies that disabling the idle timeout doesn't abort the connection if it's idle.</summary>
-    [Test]
-    public async Task Connection_with_no_idle_timeout_is_not_aborted_when_idle()
-    {
-        // Arrange
-        IServiceCollection services = CreateServiceCollection();
-
-        services.AddOptions<SlicTransportOptions>("server").Configure(
-            options => options.IdleTimeout = Timeout.InfiniteTimeSpan);
-        services.AddOptions<SlicTransportOptions>("client").Configure(
-            options => options.IdleTimeout = Timeout.InfiniteTimeSpan);
-
-        await using ServiceProvider provider = services.BuildServiceProvider(validateScopes: true);
-
-        var listener = provider.GetRequiredService<IListener<IMultiplexedConnection>>();
-        var clientTransport = provider.GetRequiredService<IMultiplexedClientTransport>();
-
-        await using var clientConnection = clientTransport.CreateConnection(
-            listener.ServerAddress,
-            provider.GetRequiredService<IOptions<MultiplexedConnectionOptions>>().Value,
-            provider.GetService<SslClientAuthenticationOptions>());
-
-        var connectTask = clientConnection.ConnectAsync(default);
-        await using var serverConnection = (await listener.AcceptAsync(default)).Connection;
-
-        _ = await serverConnection.ConnectAsync(default);
-        _ = await connectTask;
-
-        ValueTask<IMultiplexedStream> acceptTask = serverConnection.AcceptStreamAsync(default);
-
-        // Act
-        await Task.Delay(TimeSpan.FromSeconds(1));
-
-        // Assert
-        Assert.That(acceptTask.IsCompleted, Is.False);
-        await clientConnection.CloseAsync(MultiplexedConnectionCloseError.NoError, default);
-        Assert.That(async () => await acceptTask, Throws.InstanceOf<IceRpcException>());
-    }
-
-    /// <summary>Verifies that setting the idle timeout doesn't abort the connection if it's idle.</summary>
-    [Test]
-    public async Task Connection_with_idle_timeout_is_not_aborted_when_idle(
-        [Values(true, false)] bool serverIdleTimeout)
-    {
-        // Arrange
-        IServiceCollection services = CreateServiceCollection();
-
-        var idleTimeout = TimeSpan.FromSeconds(1);
-        if (serverIdleTimeout)
-        {
-            services.AddOptions<SlicTransportOptions>("server").Configure(options => options.IdleTimeout = idleTimeout);
-        }
-        else
-        {
-            services.AddOptions<SlicTransportOptions>("client").Configure(options => options.IdleTimeout = idleTimeout);
-        }
-
-        await using ServiceProvider provider = services.BuildServiceProvider(validateScopes: true);
-
-        var listener = provider.GetRequiredService<IListener<IMultiplexedConnection>>();
-        var clientTransport = provider.GetRequiredService<IMultiplexedClientTransport>();
-        await using var clientConnection = clientTransport.CreateConnection(
-            listener.ServerAddress,
-            provider.GetRequiredService<IOptions<MultiplexedConnectionOptions>>().Value,
-            provider.GetService<SslClientAuthenticationOptions>());
-
-        var connectTask = clientConnection.ConnectAsync(default);
-        await using var serverConnection = (await listener.AcceptAsync(default)).Connection;
-
-        _ = await serverConnection.ConnectAsync(default);
-        _ = await connectTask;
-
-        ValueTask<IMultiplexedStream> acceptTask = serverConnection.AcceptStreamAsync(default);
-
-        // Act
-        await Task.Delay(TimeSpan.FromSeconds(2));
-
-        // Assert
-        Assert.That(acceptTask.IsCompleted, Is.False);
-        await clientConnection.CloseAsync(MultiplexedConnectionCloseError.NoError, default);
-        Assert.That(async () => await acceptTask, Throws.InstanceOf<IceRpcException>());
-    }
-
     [Test]
     public async Task Create_client_connection_with_unknown_server_address_parameter_fails_with_format_exception()
     {
